Move GetBuildConfiguration from SBDebugger -> Debugger
